+++ cmd/av/commit.go
@@ -3,21 +3,47 @@ package main
 import (
 	"fmt"
 	"os"
+	"regexp"
+	"strings"
 
 	"emperror.dev/errors"
 	"github.com/aviator-co/av/internal/actions"
+	"github.com/aviator-co/av/internal/config"
 	"github.com/aviator-co/av/internal/git"
 	"github.com/aviator-co/av/internal/meta"
+	"github.com/aviator-co/av/internal/utils/cleanup"
 	"github.com/aviator-co/av/internal/utils/colors"
 	"github.com/shurcooL/githubv4"
+	"github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
 )
 
+const (
+	// This is an arbitrary limit on the branch name length.
+	branchNameLength = 200
+)
+
+var (
+	// See man 1 git-check-ref-format for the refname spec. This pattern is more restrictive
+	// than the spec.
+	//
+	// * Do not allow slashes because creating a branch directory from a commit message is
+	//   unlikely.
+	// * Do not allow dots because dots cannot be placed on a certain location and it's unlikely
+	//   the user wants to have a dot in the branch name.
+	branchNameReplacedPattern = regexp.MustCompile("[^-_a-zA-Z0-9]")
+
+	multipleSpacePattern = regexp.MustCompile(" +")
+)
+
 var commitFlags struct {
-	Message string
-	All     bool
-	Amend   bool
-	Edit    bool
+	Message      string
+	All          bool
+	Amend        bool
+	Edit         bool
+	CreateBranch bool
+	BranchName   string
+	AllChanges   bool
 }
 
 var commitCmd = &cobra.Command{
@@ -26,7 +52,20 @@ var commitCmd = &cobra.Command{
 	Args:  cobra.NoArgs,
 	RunE: func(cmd *cobra.Command, _ []string) error {
 		if commitFlags.Amend {
+			if commitFlags.CreateBranch || commitFlags.BranchName != "" {
+				return errors.New("cannot create a branch and amend at the same time")
+			}
 			return amendCmd(commitFlags.Message, commitFlags.Edit, commitFlags.All)
+		}
+
+		if commitFlags.CreateBranch || commitFlags.BranchName != "" {
+			return branchAndCommit(
+				commitFlags.BranchName,
+				commitFlags.Message,
+				commitFlags.AllChanges,
+				commitFlags.All,
+			)
+
 		}
 
 		repo, err := getRepo()
@@ -54,6 +93,20 @@ func runCreate(repo *git.Repo, db meta.DB) error {
 	currentBranch, err := repo.CurrentBranchName()
 	if err != nil {
 		return errors.WrapIf(err, "failed to determine current branch")
+	}
+
+	// Handle "--all-changes"
+	if commitFlags.AllChanges {
+		_, err := repo.Run(&git.RunOpts{
+			Args:      []string{"add", "--all"},
+			ExitError: true,
+		})
+		if err != nil {
+			fmt.Fprint(os.Stderr,
+				"\n", colors.Failure("Failed to stage files: ", err.Error()), "\n",
+			)
+			return actions.ErrExitSilently{ExitCode: 1}
+		}
 	}
 
 	commitArgs := []string{"commit"}
@@ -138,6 +191,20 @@ func runAmend(repo *git.Repo, db meta.DB, message string, edit bool, all bool) e
 		return errors.New("this branch has already been merged, amending is not allowed")
 	}
 
+	// Handle "--all-changes"
+	if commitFlags.AllChanges {
+		_, err := repo.Run(&git.RunOpts{
+			Args:      []string{"add", "--all"},
+			ExitError: true,
+		})
+		if err != nil {
+			fmt.Fprint(os.Stderr,
+				"\n", colors.Failure("Failed to stage files: ", err.Error()), "\n",
+			)
+			return actions.ErrExitSilently{ExitCode: 1}
+		}
+	}
+
 	if _, err := repo.Run(&git.RunOpts{
 		Args:        commitArgs,
 		ExitError:   true,
@@ -148,15 +215,179 @@ func runAmend(repo *git.Repo, db meta.DB, message string, edit bool, all bool) e
 	return nil
 }
 
+func branchAndCommit(branchName string, message string, all bool, allModified bool) (reterr error) {
+	if branchName == "" {
+		if message == "" {
+			return errors.New("Need a branch name or a commit message")
+		}
+		branchName = branchNameFromMessage(message)
+		if branchName == "" {
+			return errors.New("Cannot create a valid branch name from the message")
+		}
+	}
+
+	repo, err := getRepo()
+	if err != nil {
+		return err
+	}
+
+	db, err := getDB(repo)
+	if err != nil {
+		return err
+	}
+	tx := db.WriteTx()
+	var cu cleanup.Cleanup
+	defer cu.Cleanup()
+	cu.Add(func() {
+		logrus.WithError(reterr).Debug("aborting db transaction")
+		tx.Abort()
+	})
+
+	parentBranchName, err := repo.CurrentBranchName()
+	if err != nil {
+		return errors.WrapIff(err, "failed to get current branch name")
+	}
+
+	// Currently, we only allow the repo default branch to be a trunk.
+	// We might want to allow other branches to be trunks in the future, but
+	// that does run the risk of allowing the user to get into a weird state
+	// (where some stacks assume a branch is a trunk and others don't).
+	isBranchFromTrunk, err := repo.IsTrunkBranch(parentBranchName)
+	if err != nil {
+		return errors.WrapIf(err, "failed to check if branch is trunk")
+	}
+	var parentHead string
+	if !isBranchFromTrunk {
+		parentHead, err = repo.RevParse(&git.RevParse{Rev: parentBranchName})
+		if err != nil {
+			return errors.WrapIf(err, "failed to get parent branch head commit")
+		}
+	}
+
+	if err != nil {
+		return errors.WrapIf(err, "failed to read parent branch state")
+	}
+
+	// Create a new branch off of the parent
+	logrus.WithFields(logrus.Fields{
+		"parent":     parentBranchName,
+		"new_branch": branchName,
+	}).Debug("creating new branch from parent")
+	if _, err := repo.CheckoutBranch(&git.CheckoutBranch{
+		Name:      branchName,
+		NewBranch: true,
+	}); err != nil {
+		return errors.WrapIff(err, "checkout error")
+	}
+
+	// On failure, we want to delete the branch we created so that the user
+	// can try again (e.g., to fix issues surfaced by a pre-commit hook).
+	cu.Add(func() {
+		fmt.Fprint(os.Stderr,
+			colors.Faint("  - Cleaning up branch "),
+			colors.UserInput(branchName),
+			colors.Faint(" because commit was not successful."),
+			"\n",
+		)
+		if _, err := repo.CheckoutBranch(&git.CheckoutBranch{
+			Name: parentBranchName,
+		}); err != nil {
+			logrus.WithError(err).Error("failed to return to original branch during cleanup")
+		}
+		if err := repo.BranchDelete(branchName); err != nil {
+			logrus.WithError(err).Error("failed to delete branch during cleanup")
+		}
+	})
+
+	tx.SetBranch(meta.Branch{
+		Name: branchName,
+		Parent: meta.BranchState{
+			Name:  parentBranchName,
+			Trunk: isBranchFromTrunk,
+			Head:  parentHead,
+		},
+	})
+
+	// For "--all" and "--all-modified",
+	var addArgs []string
+	if all {
+		addArgs = append(addArgs, "--all")
+	} else if allModified {
+		// This is meant to mirror `git commit --all` which does not add
+		// unstaged files to the index (which is different from `git add --all`).
+		addArgs = append(addArgs, "--update")
+	}
+	if len(addArgs) > 0 {
+		_, err := repo.Run(&git.RunOpts{
+			Args:      append([]string{"add"}, addArgs...),
+			ExitError: true,
+		})
+		if err != nil {
+			fmt.Fprint(os.Stderr,
+				"\n", colors.Failure("Failed to stage files: ", err.Error()), "\n",
+			)
+			return actions.ErrExitSilently{ExitCode: 1}
+		}
+	}
+
+	commitArgs := []string{"commit"}
+	if message != "" {
+		commitArgs = append(commitArgs, "--message", message)
+	}
+
+	if _, err := repo.Run(&git.RunOpts{
+		Args:        commitArgs,
+		ExitError:   true,
+		Interactive: true,
+	}); err != nil {
+		fmt.Fprint(os.Stderr,
+			"\n", colors.Failure("Failed to create commit."), "\n",
+		)
+		return actions.ErrExitSilently{ExitCode: 1}
+	}
+
+	// Cancel the cleanup **after** the commit is successful (so that we
+	// delete the created branch).
+	cu.Cancel()
+	if err := tx.Commit(); err != nil {
+		return err
+	}
+
+	return nil
+}
+
+func branchNameFromMessage(message string) string {
+	name := branchNameReplacedPattern.ReplaceAllLiteralString(message, " ")
+	name = strings.TrimSpace(name)
+	name = multipleSpacePattern.ReplaceAllLiteralString(name, "-")
+	if len(name) > branchNameLength {
+		name = name[:branchNameLength]
+	}
+	name = strings.ToLower(name)
+	if config.Av.PullRequest.BranchNamePrefix != "" {
+		name = fmt.Sprintf("%s%s", config.Av.PullRequest.BranchNamePrefix, name)
+	}
+	return name
+}
+
 func init() {
 	commitCmd.Flags().
 		StringVarP(&commitFlags.Message, "message", "m", "", "the commit message")
 	commitCmd.Flags().
-		BoolVarP(&commitFlags.All, "all", "a", false, "automatically stage modified files (same as git commit --all)")
+		BoolVarP(&commitFlags.All, "all", "a", false, "automatically stage modified and deleted files (same as git commit --all)")
 	commitCmd.Flags().
-		BoolVarP(&commitFlags.Amend, "amend", "", false, "amend the last commit")
+		BoolVarP(&commitFlags.AllChanges, "all-changes", "A", false, "automatically stage all files, including untracked files")
 	commitCmd.Flags().
-		BoolVarP(&commitFlags.Edit, "edit", "", false, "edit an amended commit's message")
+		BoolVar(&commitFlags.Amend, "amend", false, "amend the last commit")
+	commitCmd.Flags().
+		BoolVar(&commitFlags.Edit, "edit", false, "edit an amended commit's message")
+	commitCmd.Flags().
+		BoolVarP(&commitFlags.CreateBranch, "branch", "b", false,
+			"create a new branch with automatically generated name and commit to it")
+	commitCmd.Flags().
+		StringVar(&commitFlags.BranchName, "branch-name", "", "create a new branch with the given name and commit to it")
+
+	commitCmd.MarkFlagsMutuallyExclusive("all", "all-changes")
 
 	deprecatedAmendCmd := deprecateCommand(*commitAmendCmd, "av commit --amend", "amend")
 	deprecatedAmendCmd.Hidden = true
@@ -178,5 +409,4 @@ func init() {
 		deprecatedCreateCmd,
 		deprecatedSplitCmd,
 	)
-
 }
